@@ -25,13 +25,14 @@
     "js-cookie": "^3.0.5",
     "jspdf": "^2.5.2",
     "jspdf-autotable": "^3.8.4",
+    "lucide-react": "^0.487.0",
     "mime-types": "^2.1.35",
-    "next": "^15.1.7",
+    "next": "^15.3.0",
     "next-pwa": "^5.6.0",
     "nodemailer": "^6.10.0",
     "prop-types": "^15.8.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-hook-form": "^7.54.2",
     "react-icons": "^5.4.0",
     "react-intersection-observer": "^9.15.1",
