@@ -1,7 +1,8 @@
 "use client"
 
-import { useState, useRef, useMemo, useEffect } from "react"
-import PropTypes from "prop-types"
+import { useState, useMemo, useEffect } from "react"
+import { useRouter } from "next/router"
+import Head from "next/head"
 import { doc, addDoc, collection, getDoc, updateDoc, setDoc } from "firebase/firestore"
 import { createUserWithEmailAndPassword, sendEmailVerification } from "firebase/auth"
 import { auth, db } from "../lib/firebase"
@@ -9,7 +10,10 @@ import { toast } from "sonner"
 import { jsPDF } from "jspdf"
 import "jspdf-autotable"
 import citiesData from "../data/cities.json"
-import { Check, ChevronRight, ChevronLeft, X } from "lucide-react"
+import { Check, ChevronRight, ChevronLeft, ArrowLeft } from "lucide-react"
+import Layout from "../components/Layout"
+import { useBasket } from "../contexts/BasketContext"
+import Link from "next/link"
 
 // PDF generation helper function
 const generateInvoice = (order, userData) => {
@@ -222,16 +226,6 @@ const FormInput = ({ label, name, value, onChange, type = "text", required = tru
   </div>
 )
 
-FormInput.propTypes = {
-  label: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
-  onChange: PropTypes.func.isRequired,
-  type: PropTypes.string,
-  required: PropTypes.bool,
-  children: PropTypes.node,
-}
-
 // FormSelect Component for dropdowns
 const FormSelect = ({ label, name, value, onChange, options, required = true, ...props }) => (
   <div className="mb-6">
@@ -257,22 +251,10 @@ const FormSelect = ({ label, name, value, onChange, options, required = true, ..
   </div>
 )
 
-FormSelect.propTypes = {
-  label: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
-  onChange: PropTypes.func.isRequired,
-  options: PropTypes.arrayOf(
-    PropTypes.shape({
-      value: PropTypes.string.isRequired,
-      label: PropTypes.string.isRequired,
-    }),
-  ).isRequired,
-  required: PropTypes.bool,
-}
+export default function CheckoutPage() {
+  const router = useRouter()
+  const { basketItems, basketCount, clearBasket } = useBasket()
 
-// CheckoutPopup Component
-export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
   // Checkout steps
   const STEPS = {
     SHIPPING: "shipping",
@@ -297,12 +279,14 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
     password: "",
     confirmPassword: "",
     createAccount: false,
+    gender: "", // Added gender field
   })
 
   const [loading, setLoading] = useState(false)
   const [placedOrder, setPlacedOrder] = useState(null)
   const [userDataForInvoice, setUserDataForInvoice] = useState(null)
   const [isLoggedIn, setIsLoggedIn] = useState(false)
+  const [isFromBasket, setIsFromBasket] = useState(false)
 
   // State for location data options
   const [locationOptions, setLocationOptions] = useState({
@@ -311,12 +295,26 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
     districts: [],
   })
 
-  const popupRef = useRef(null)
-
   const subtotal = useMemo(() => basketItems.reduce((sum, item) => sum + item.price * item.quantity, 0), [basketItems])
-
   const shipping = subtotal > 200 ? 0 : 8.0 // Free shipping over 200 TND
   const totalAmount = subtotal + shipping
+
+  // Check if user came from basket and has items
+  useEffect(() => {
+    // Check if the user came from the basket page
+    const referrer = document.referrer
+    const fromBasket = referrer.includes("/basket") || sessionStorage.getItem("fromBasket") === "true"
+    setIsFromBasket(fromBasket)
+
+    // Store that we're in the checkout flow
+    sessionStorage.setItem("fromBasket", "true")
+
+    // If basket is empty or user didn't come from basket, redirect to basket
+    if (basketCount === 0) {
+      toast.error("Your basket is empty. Please add items before checkout.")
+      router.push("/basket")
+    }
+  }, [basketCount, router])
 
   // Check if user is logged in
   useEffect(() => {
@@ -505,6 +503,11 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
       return false
     }
 
+    if (!accountData.gender) {
+      toast.error("Please select your gender")
+      return false
+    }
+
     return true
   }
 
@@ -540,32 +543,84 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
     if (!accountData.createAccount) return null
 
     try {
-      // Create user with email and password
-      const userCredential = await createUserWithEmailAndPassword(auth, accountData.email, accountData.password)
+      // Validate email format
+      if (!/\S+@\S+\.\S+/.test(shippingData.email)) {
+        toast.error("Please enter a valid email address")
+        return null
+      }
 
-      // Send email verification
-      await sendEmailVerification(userCredential.user)
+      // Validate phone format
+      if (!/^\d{8,15}$/.test(shippingData.phone.replace(/\D/g, ""))) {
+        toast.error("Phone number must be 8-15 digits")
+        return null
+      }
 
-      // Create user document in Firestore
-      await setDoc(doc(db, "users", userCredential.user.uid), {
+      // Create user with Firebase Authentication
+      const userCredential = await createUserWithEmailAndPassword(auth, shippingData.email, accountData.password)
+      const user = userCredential.user
+
+      // Save user details in Firestore
+      await setDoc(doc(db, "users", user.uid), {
         firstName: shippingData.firstName,
         lastName: shippingData.lastName,
-        email: shippingData.email,
         phone: shippingData.phone,
+        email: shippingData.email,
+        gender: accountData.gender,
         createdAt: new Date().toISOString(),
+        role: "user",
       })
 
-      toast.success("Account created successfully! Please check your email for verification.")
-      return userCredential.user
+      // Add user to newsletter_signups collection
+      await addDoc(collection(db, "newsletter_signups"), {
+        email: shippingData.email,
+        timestamp: new Date().toISOString(),
+      })
+
+      // Create an empty basket for the user
+      await setDoc(doc(db, "baskets", user.uid), {
+        items: [],
+      })
+
+      // Try to send verification email via backend
+      try {
+        await fetch("https://auth.ashe.tn/auth/verify-email", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ email: shippingData.email }),
+        })
+      } catch (emailError) {
+        console.error("Failed to send verification email:", emailError)
+        // Continue with account creation even if email sending fails
+      }
+
+      // Send email verification through Firebase as fallback
+      await sendEmailVerification(user)
+
+      toast.success("Account created! Please check your email to verify your account.")
+      return user
     } catch (error) {
       console.error("Error creating account:", error)
-      toast.error(error.message || "Failed to create account")
+
+      // Handle specific Firebase auth errors
+      const errorMap = {
+        "auth/email-already-in-use": "This email is already in use. Please use a different email.",
+        "auth/weak-password": "Password is too weak. Please use a stronger password.",
+        "auth/invalid-email": "Please enter a valid email address.",
+        "auth/network-request-failed": "Network error. Please check your connection.",
+      }
+
+      toast.error(errorMap[error.code] || "Failed to create account. Please try again.")
       return null
     }
   }
 
   const handlePlaceOrder = async () => {
     setLoading(true)
+
+    // Reset any previous order state to ensure clean rendering
+    setPlacedOrder(null)
+    setUserDataForInvoice(null)
+
     try {
       let user = auth.currentUser
 
@@ -661,15 +716,27 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
       )
 
       // Save order details in state
-      setPlacedOrder(orderWithId)
+      console.log("Order placed successfully:", orderWithId)
+
+      // Save order details in state - use a callback to ensure state is updated
+      setPlacedOrder(() => orderWithId)
       setUserDataForInvoice({
         firstName: shippingData.firstName,
         lastName: shippingData.lastName,
         phone: shippingData.phone,
       })
 
-      onPlaceOrder(orderWithId)
+      // Show success message before clearing basket
       toast.success("Order placed successfully!")
+
+      // Add a small delay before clearing the basket to ensure state updates are processed
+      setTimeout(() => {
+        // Clear basket after successful order
+        clearBasket()
+
+        // Clear the fromBasket flag
+        sessionStorage.removeItem("fromBasket")
+      }, 300)
     } catch (error) {
       console.error("Order error:", error)
       toast.error(error.message || "Failed to place order. Please try again.")
@@ -678,53 +745,16 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
     }
   }
 
-  // If order has been placed, show the confirmation and Download Invoice button
-  if (placedOrder) {
-    return (
-      <div className="fixed inset-0 bg-white/90 backdrop-blur-sm flex flex-col justify-center items-center z-50 p-4">
-        <div className="bg-white w-full max-w-md p-8 shadow-xl rounded-lg">
-          <div className="text-center mb-6">
-            <div className="w-16 h-16 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-4">
-              <Check className="w-8 h-8 text-green-600" />
-            </div>
-            <h2 className="text-2xl font-bold mb-2">Order Confirmed!</h2>
-            <p className="text-gray-600">
-              Your order <span className="font-semibold">#{placedOrder.id}</span> has been placed successfully.
-            </p>
-          </div>
-
-          <div className="bg-gray-50 p-4 rounded-lg mb-6">
-            <div className="flex justify-between mb-2">
-              <span className="text-gray-600">Subtotal:</span>
-              <span>{subtotal.toFixed(2)} TND</span>
-            </div>
-            <div className="flex justify-between mb-2">
-              <span className="text-gray-600">Shipping:</span>
-              <span>{shipping === 0 ? "Free" : `${shipping.toFixed(2)} TND`}</span>
-            </div>
-            <div className="flex justify-between font-bold">
-              <span>Total:</span>
-              <span>{totalAmount.toFixed(2)} TND</span>
-            </div>
-          </div>
-
-          <button
-            onClick={() => generateInvoice(placedOrder, userDataForInvoice)}
-            className="w-full py-3 mb-4 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center justify-center gap-2"
-          >
-            <span>Download Invoice</span>
-          </button>
-
-          <button
-            onClick={onClose}
-            className="w-full py-3 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors"
-          >
-            Continue Shopping
-          </button>
-        </div>
-      </div>
-    )
-  }
+  // Monitor placedOrder state changes
+  useEffect(() => {
+    if (placedOrder) {
+      console.log("Order placed, confirmation page should render:", placedOrder.id)
+      // Force a re-render if needed
+      setTimeout(() => {
+        window.scrollTo(0, 0)
+      }, 100)
+    }
+  }, [placedOrder])
 
   // Order Summary Component
   const OrderSummary = () => (
@@ -808,279 +838,407 @@ export default function CheckoutPopup({ basketItems, onClose, onPlaceOrder }) {
     )
   }
 
-  return (
-    <div className="fixed inset-0 bg-black/50 backdrop-blur-sm flex justify-center items-center z-50 p-4">
-      <div
-        ref={popupRef}
-        className="bg-white w-full max-w-6xl h-[90vh] flex flex-col lg:grid lg:grid-cols-5 shadow-xl rounded-lg overflow-hidden"
-      >
-        {/* Left Panel - Order Summary */}
-        <div className="lg:col-span-2 bg-gray-50 p-6 overflow-y-auto">
-          <button
-            onClick={onClose}
-            className="absolute top-4 left-4 text-gray-500 hover:text-gray-700 p-2 rounded-full transition-all"
-            aria-label="Close checkout"
-            disabled={loading}
-          >
-            <X className="w-5 h-5" />
-          </button>
-
-          <div className="mt-8 mb-6">
-            <h1 className="text-2xl font-bold">ASHE™</h1>
-            <p className="text-gray-500">Secure Checkout</p>
-          </div>
-
-          <OrderSummary />
-
-          <div className="mt-6 space-y-4">
-            <div className="flex items-center gap-2 text-sm text-gray-600">
-              <Check className="w-4 h-4 text-green-500" />
-              <span>Secure payment</span>
+  // If order has been placed, show the confirmation and Download Invoice button
+  if (placedOrder !== null) {
+    console.log("Rendering order confirmation page for order:", placedOrder.id)
+    return (
+      <Layout>
+        <Head>
+          <title>Order Confirmation | ASHE™</title>
+          <meta name="description" content="Your order has been confirmed. Thank you for shopping with ASHE™." />
+        </Head>
+        <div className="max-w-4xl mx-auto px-4 py-16">
+          <div className="bg-white p-8 shadow-xl rounded-lg">
+            <div className="text-center mb-8">
+              <div className="w-20 h-20 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-4">
+                <Check className="w-10 h-10 text-green-600" />
+              </div>
+              <h1 className="text-3xl font-bold mb-2">Order Confirmed!</h1>
+              <p className="text-gray-600 text-lg">
+                Your order <span className="font-semibold">#{placedOrder.id}</span> has been placed successfully.
+              </p>
             </div>
-            <div className="flex items-center gap-2 text-sm text-gray-600">
-              <Check className="w-4 h-4 text-green-500" />
-              <span>Free shipping on orders over 200 TND</span>
+
+            <div className="bg-gray-50 p-6 rounded-lg mb-8">
+              <h2 className="text-xl font-bold mb-4">Order Summary</h2>
+              <div className="space-y-4">
+                <div className="flex justify-between mb-2">
+                  <span className="text-gray-600">Subtotal:</span>
+                  <span>{subtotal.toFixed(2)} TND</span>
+                </div>
+                <div className="flex justify-between mb-2">
+                  <span className="text-gray-600">Shipping:</span>
+                  <span>{shipping === 0 ? "Free" : `${shipping.toFixed(2)} TND`}</span>
+                </div>
+                <div className="border-t pt-4 flex justify-between font-bold text-lg">
+                  <span>Total:</span>
+                  <span>{totalAmount.toFixed(2)} TND</span>
+                </div>
+              </div>
             </div>
-            <div className="flex items-center gap-2 text-sm text-gray-600">
-              <Check className="w-4 h-4 text-green-500" />
-              <span>7-day return policy</span>
+
+            <div className="bg-gray-50 p-6 rounded-lg mb-8">
+              <h2 className="text-xl font-bold mb-4">Shipping Information</h2>
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                <div>
+                  <h3 className="font-medium text-gray-700">Contact</h3>
+                  <p>{placedOrder.userInfo.name}</p>
+                  <p>{placedOrder.userInfo.email}</p>
+                  <p>{placedOrder.userInfo.phone}</p>
+                </div>
+                <div>
+                  <h3 className="font-medium text-gray-700">Delivery Address</h3>
+                  <p>{placedOrder.shippingInfo.addressLine}</p>
+                  <p>
+                    {placedOrder.shippingInfo.district}, {placedOrder.shippingInfo.delegation}
+                  </p>
+                  <p>{placedOrder.shippingInfo.governorate}</p>
+                </div>
+              </div>
+            </div>
+
+            <div className="flex flex-col sm:flex-row gap-4 justify-center">
+              <button
+                onClick={() => generateInvoice(placedOrder, userDataForInvoice)}
+                className="py-3 px-6 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center justify-center gap-2"
+              >
+                <span>Download Invoice</span>
+              </button>
+
+              <Link href="/products">
+                <a className="py-3 px-6 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors flex items-center justify-center">
+                  Continue Shopping
+                </a>
+              </Link>
             </div>
           </div>
         </div>
+      </Layout>
+    )
+  }
 
-        {/* Right Panel - Checkout Form */}
-        <div className="lg:col-span-3 p-6 overflow-y-auto">
-          <ProgressIndicator />
+  return (
+    <Layout>
+      <Head>
+        <title>Checkout | ASHE™</title>
+        <meta name="description" content="Complete your purchase securely with ASHE™." />
+      </Head>
 
-          {/* Shipping Information Step */}
-          {currentStep === STEPS.SHIPPING && (
-            <div>
-              <h2 className="text-xl font-bold mb-6">Shipping Information</h2>
-              <div className="grid grid-cols-2 gap-4">
-                <FormInput
-                  label="First Name"
-                  name="firstName"
-                  value={shippingData.firstName}
-                  onChange={handleShippingChange}
-                  placeholder="John"
-                />
-                <FormInput
-                  label="Last Name"
-                  name="lastName"
-                  value={shippingData.lastName}
-                  onChange={handleShippingChange}
-                  placeholder="Doe"
-                />
-              </div>
+      <div className="max-w-7xl mx-auto px-4 py-12">
+        <div className="mb-6">
+          <Link href="/basket">
+            <a className="inline-flex items-center text-gray-600 hover:text-[#46c7c7] transition-colors">
+              <ArrowLeft className="w-4 h-4 mr-2" />
+              Back to Basket
+            </a>
+          </Link>
+        </div>
 
-              <FormInput
-                label="Email"
-                name="email"
-                type="email"
-                value={shippingData.email}
-                onChange={handleShippingChange}
-                placeholder="your@email.com"
-              />
+        <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
+          {/* Left Column - Checkout Form */}
+          <div className="lg:col-span-2">
+            <div className="bg-white rounded-lg shadow-sm p-6">
+              <h1 className="text-2xl font-bold mb-6">Checkout</h1>
 
-              <FormInput
-                label="Phone"
-                name="phone"
-                value={shippingData.phone}
-                onChange={handleShippingChange}
-                placeholder="e.g., 20123456"
-              />
+              <ProgressIndicator />
 
-              <FormSelect
-                label="Governorate"
-                name="governorate"
-                value={shippingData.governorate}
-                onChange={handleShippingChange}
-                options={locationOptions.governorates}
-              />
+              {/* Shipping Information Step */}
+              {currentStep === STEPS.SHIPPING && (
+                <div>
+                  <h2 className="text-xl font-bold mb-6">Shipping Information</h2>
+                  <div className="grid grid-cols-2 gap-4">
+                    <FormInput
+                      label="First Name"
+                      name="firstName"
+                      value={shippingData.firstName}
+                      onChange={handleShippingChange}
+                      placeholder="John"
+                    />
+                    <FormInput
+                      label="Last Name"
+                      name="lastName"
+                      value={shippingData.lastName}
+                      onChange={handleShippingChange}
+                      placeholder="Doe"
+                    />
+                  </div>
 
-              <div className="grid grid-cols-2 gap-4">
-                <FormSelect
-                  label="Delegation"
-                  name="delegation"
-                  value={shippingData.delegation}
-                  onChange={handleShippingChange}
-                  options={locationOptions.delegations}
-                  disabled={!shippingData.governorate}
-                />
-
-                <FormSelect
-                  label="District"
-                  name="district"
-                  value={shippingData.district}
-                  onChange={handleShippingChange}
-                  options={locationOptions.districts}
-                  disabled={!shippingData.delegation}
-                />
-              </div>
-
-              <FormInput
-                label="Address Line"
-                name="addressLine"
-                value={shippingData.addressLine}
-                onChange={handleShippingChange}
-                placeholder="Street name, building number, apartment, etc."
-              />
-            </div>
-          )}
-
-          {/* Account Creation Step (only for non-logged in users) */}
-          {currentStep === STEPS.ACCOUNT && (
-            <div>
-              <h2 className="text-xl font-bold mb-6">Create an Account (Optional)</h2>
-
-              <div className="mb-6">
-                <div className="flex items-center mb-4">
-                  <input
-                    id="createAccount"
-                    name="createAccount"
-                    type="checkbox"
-                    checked={accountData.createAccount}
-                    onChange={handleAccountChange}
-                    className="w-4 h-4 text-[#46c7c7] border-gray-300 rounded focus:ring-[#46c7c7]"
+                  <FormInput
+                    label="Email"
+                    name="email"
+                    type="email"
+                    value={shippingData.email}
+                    onChange={handleShippingChange}
+                    placeholder="your@email.com"
                   />
-                  <label htmlFor="createAccount" className="ml-2 text-sm font-medium text-gray-700">
-                    Create an account for faster checkout next time
-                  </label>
-                </div>
 
-                {accountData.createAccount && (
-                  <div className="space-y-4 p-4 border border-gray-200 rounded-lg bg-gray-50">
-                    <p className="text-sm text-gray-600 mb-4">
-                      Your account will be created with the email: <strong>{shippingData.email}</strong>
-                    </p>
+                  <FormInput
+                    label="Phone"
+                    name="phone"
+                    value={shippingData.phone}
+                    onChange={handleShippingChange}
+                    placeholder="e.g., 20123456"
+                  />
 
-                    <FormInput
-                      label="Password"
-                      name="password"
-                      type="password"
-                      value={accountData.password}
-                      onChange={handleAccountChange}
-                      placeholder="Create a password"
+                  <FormSelect
+                    label="Governorate"
+                    name="governorate"
+                    value={shippingData.governorate}
+                    onChange={handleShippingChange}
+                    options={locationOptions.governorates}
+                  />
+
+                  <div className="grid grid-cols-2 gap-4">
+                    <FormSelect
+                      label="Delegation"
+                      name="delegation"
+                      value={shippingData.delegation}
+                      onChange={handleShippingChange}
+                      options={locationOptions.delegations}
+                      disabled={!shippingData.governorate}
                     />
 
-                    <FormInput
-                      label="Confirm Password"
-                      name="confirmPassword"
-                      type="password"
-                      value={accountData.confirmPassword}
-                      onChange={handleAccountChange}
-                      placeholder="Confirm your password"
+                    <FormSelect
+                      label="District"
+                      name="district"
+                      value={shippingData.district}
+                      onChange={handleShippingChange}
+                      options={locationOptions.districts}
+                      disabled={!shippingData.delegation}
                     />
                   </div>
-                )}
-              </div>
-            </div>
-          )}
 
-          {/* Order Confirmation Step */}
-          {currentStep === STEPS.CONFIRMATION && (
-            <div>
-              <h2 className="text-xl font-bold mb-6">Confirm Your Order</h2>
-
-              <div className="space-y-6">
-                <div className="bg-gray-50 p-4 rounded-lg">
-                  <h3 className="font-medium mb-2">Shipping Information</h3>
-                  <p>
-                    {shippingData.firstName} {shippingData.lastName}
-                  </p>
-                  <p>{shippingData.addressLine}</p>
-                  <p>
-                    {shippingData.district}, {shippingData.delegation}
-                  </p>
-                  <p>{shippingData.governorate}</p>
-                  <p>Phone: {shippingData.phone}</p>
-                  <p>Email: {shippingData.email}</p>
+                  <FormInput
+                    label="Address Line"
+                    name="addressLine"
+                    value={shippingData.addressLine}
+                    onChange={handleShippingChange}
+                    placeholder="Street name, building number, apartment, etc."
+                  />
                 </div>
+              )}
 
-                <div className="bg-gray-50 p-4 rounded-lg">
-                  <h3 className="font-medium mb-2">Payment Method</h3>
-                  <p>Cash on Delivery</p>
-                  <p className="text-sm text-gray-600 mt-1">You will pay when your order is delivered.</p>
-                </div>
+              {/* Account Creation Step (only for non-logged in users) */}
+              {currentStep === STEPS.ACCOUNT && (
+                <div>
+                  <h2 className="text-xl font-bold mb-6">Create an Account (Optional)</h2>
 
-                {!isLoggedIn && accountData.createAccount && (
-                  <div className="bg-gray-50 p-4 rounded-lg">
-                    <h3 className="font-medium mb-2">Account</h3>
-                    <p>An account will be created with your email address</p>
+                  <div className="mb-6">
+                    <div className="flex items-center mb-4">
+                      <input
+                        id="createAccount"
+                        name="createAccount"
+                        type="checkbox"
+                        checked={accountData.createAccount}
+                        onChange={handleAccountChange}
+                        className="w-4 h-4 text-[#46c7c7] border-gray-300 rounded focus:ring-[#46c7c7]"
+                      />
+                      <label htmlFor="createAccount" className="ml-2 text-sm font-medium text-gray-700">
+                        Create an account for faster checkout next time
+                      </label>
+                    </div>
+
+                    {accountData.createAccount && (
+                      <div className="space-y-4 p-4 border border-gray-200 rounded-lg bg-gray-50">
+                        <p className="text-sm text-gray-600 mb-4">
+                          Your account will be created with the email: <strong>{shippingData.email}</strong>
+                        </p>
+
+                        <FormInput
+                          label="Password"
+                          name="password"
+                          type="password"
+                          value={accountData.password}
+                          onChange={handleAccountChange}
+                          placeholder="Create a password"
+                        />
+
+                        <FormInput
+                          label="Confirm Password"
+                          name="confirmPassword"
+                          type="password"
+                          value={accountData.confirmPassword}
+                          onChange={handleAccountChange}
+                          placeholder="Confirm your password"
+                        />
+
+                        {/* Gender Selection */}
+                        <div className="mb-6">
+                          <label className="block text-sm font-medium text-gray-700 mb-2">
+                            Gender <span className="text-red-500">*</span>
+                          </label>
+                          <div className="flex gap-4">
+                            {["male", "female"].map((option) => (
+                              <label key={option} className="flex items-center gap-2 cursor-pointer group">
+                                <input
+                                  type="radio"
+                                  name="gender"
+                                  value={option}
+                                  checked={accountData.gender === option}
+                                  onChange={(e) => setAccountData((prev) => ({ ...prev, gender: e.target.value }))}
+                                  className="peer hidden"
+                                  required={accountData.createAccount}
+                                />
+                                <div
+                                  className="w-4 h-4 rounded-full border-2 border-gray-300 flex items-center justify-center
+                                  group-hover:border-[#46c7c7] transition-colors duration-200
+                                  peer-checked:border-[#46c7c7] peer-checked:bg-[#46c7c7]
+                                  peer-focus-visible:ring-2 peer-focus-visible:ring-[#46c7c7]"
+                                >
+                                  <div className="w-1.5 h-1.5 rounded-full bg-white opacity-0 peer-checked:opacity-100 transition-opacity" />
+                                </div>
+                                <span className="text-sm text-gray-700 group-hover:text-gray-900 capitalize transition-colors">
+                                  {option}
+                                </span>
+                              </label>
+                            ))}
+                          </div>
+                        </div>
+                      </div>
+                    )}
                   </div>
-                )}
-              </div>
-            </div>
-          )}
+                </div>
+              )}
 
-          {/* Navigation Buttons */}
-          <div className="flex justify-between mt-8">
-            {currentStep !== STEPS.SHIPPING ? (
-              <button
-                type="button"
-                onClick={handlePrevStep}
-                className="px-6 py-3 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors flex items-center gap-2"
-                disabled={loading}
-              >
-                <ChevronLeft className="w-4 h-4" />
-                <span>Back</span>
-              </button>
-            ) : (
-              <button
-                type="button"
-                onClick={onClose}
-                className="px-6 py-3 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors"
-                disabled={loading}
-              >
-                Cancel
-              </button>
-            )}
+              {/* Order Confirmation Step */}
+              {currentStep === STEPS.CONFIRMATION && (
+                <div>
+                  <h2 className="text-xl font-bold mb-6">Confirm Your Order</h2>
 
-            {currentStep !== STEPS.CONFIRMATION ? (
-              <button
-                type="button"
-                onClick={handleNextStep}
-                className="px-6 py-3 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center gap-2"
-                disabled={loading}
-              >
-                <span>Continue</span>
-                <ChevronRight className="w-4 h-4" />
-              </button>
-            ) : (
-              <button
-                type="button"
-                onClick={handlePlaceOrder}
-                className="px-6 py-3 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center gap-2"
-                disabled={loading}
-              >
-                {loading ? (
-                  <span>Processing...</span>
+                  <div className="space-y-6">
+                    <div className="bg-gray-50 p-4 rounded-lg">
+                      <h3 className="font-medium mb-2">Shipping Information</h3>
+                      <p>
+                        {shippingData.firstName} {shippingData.lastName}
+                      </p>
+                      <p>{shippingData.addressLine}</p>
+                      <p>
+                        {shippingData.district}, {shippingData.delegation}
+                      </p>
+                      <p>{shippingData.governorate}</p>
+                      <p>Phone: {shippingData.phone}</p>
+                      <p>Email: {shippingData.email}</p>
+                    </div>
+
+                    <div className="bg-gray-50 p-4 rounded-lg">
+                      <h3 className="font-medium mb-2">Payment Method</h3>
+                      <p>Cash on Delivery</p>
+                      <p className="text-sm text-gray-600 mt-1">You will pay when your order is delivered.</p>
+                    </div>
+
+                    {!isLoggedIn && accountData.createAccount && (
+                      <div className="bg-gray-50 p-4 rounded-lg">
+                        <h3 className="font-medium mb-2">Account</h3>
+                        <p>An account will be created with your email address</p>
+                        {accountData.gender && (
+                          <p>Gender: {accountData.gender.charAt(0).toUpperCase() + accountData.gender.slice(1)}</p>
+                        )}
+                      </div>
+                    )}
+                  </div>
+                </div>
+              )}
+
+              {/* Navigation Buttons */}
+              <div className="flex justify-between mt-8">
+                {currentStep !== STEPS.SHIPPING ? (
+                  <button
+                    type="button"
+                    onClick={handlePrevStep}
+                    className="px-6 py-3 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors flex items-center gap-2"
+                    disabled={loading}
+                  >
+                    <ChevronLeft className="w-4 h-4" />
+                    <span>Back</span>
+                  </button>
                 ) : (
-                  <>
-                    <span>Place Order</span>
-                    <Check className="w-4 h-4" />
-                  </>
+                  <Link href="/basket">
+                    <a className="px-6 py-3 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors">
+                      Cancel
+                    </a>
+                  </Link>
                 )}
-              </button>
-            )}
+
+                {currentStep !== STEPS.CONFIRMATION ? (
+                  <button
+                    type="button"
+                    onClick={handleNextStep}
+                    className="px-6 py-3 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center gap-2"
+                    disabled={loading}
+                  >
+                    <span>Continue</span>
+                    <ChevronRight className="w-4 h-4" />
+                  </button>
+                ) : (
+                  <button
+                    type="button"
+                    onClick={handlePlaceOrder}
+                    className="px-6 py-3 bg-black text-white rounded-lg hover:bg-[#46c7c7] transition-colors flex items-center gap-2"
+                    disabled={loading}
+                  >
+                    {loading ? (
+                      <span>Processing...</span>
+                    ) : (
+                      <>
+                        <span>Place Order</span>
+                        <Check className="w-4 h-4" />
+                      </>
+                    )}
+                  </button>
+                )}
+              </div>
+            </div>
+          </div>
+
+          {/* Right Column - Order Summary */}
+          <div className="lg:col-span-1">
+            <div className="bg-white rounded-lg shadow-sm sticky top-6">
+              <div className="p-6 border-b">
+                <h2 className="text-lg font-medium text-gray-900">Order Summary</h2>
+              </div>
+
+              <div className="p-6">
+                <OrderSummary />
+
+                <div className="mt-6 space-y-4">
+                  <div className="flex items-center gap-2 text-sm text-gray-600">
+                    <Check className="w-4 h-4 text-green-500" />
+                    <span>Secure checkout</span>
+                  </div>
+                  <div className="flex items-center gap-2 text-sm text-gray-600">
+                    <Check className="w-4 h-4 text-green-500" />
+                    <span>Free shipping on orders over 200 TND</span>
+                  </div>
+                  <div className="flex items-center gap-2 text-sm text-gray-600">
+                    <Check className="w-4 h-4 text-green-500" />
+                    <span>7-day return policy</span>
+                  </div>
+                </div>
+              </div>
+            </div>
           </div>
         </div>
       </div>
-    </div>
+    </Layout>
   )
 }
 
-CheckoutPopup.propTypes = {
-  basketItems: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-      price: PropTypes.number.isRequired,
-      quantity: PropTypes.number.isRequired,
-      size: PropTypes.string.isRequired,
-      color: PropTypes.string.isRequired,
-    }),
-  ).isRequired,
-  onClose: PropTypes.func.isRequired,
-  onPlaceOrder: PropTypes.func.isRequired,
+export async function getServerSideProps(context) {
+  // Check if the user is coming from the basket page
+  const referer = context.req.headers.referer || ""
+  const isFromBasket = referer.includes("/basket")
+
+  if (!isFromBasket) {
+    // If not coming from basket, redirect to basket page
+    return {
+      redirect: {
+        destination: "/basket",
+        permanent: false,
+      },
+    }
+  }
+
+  return {
+    props: {}, // will be passed to the page component as props
+  }
 }
